Move `Operation` methods into `OperationBase.cpp`
