+++ src/IR/Operation.cpp
@@ -5,8 +5,6 @@
 #include "caffeine/IR/Value.h"
 #include "caffeine/Support/LLVMFmt.h"
 #include "caffeine/Support/Macros.h"
-
-#include <boost/algorithm/string.hpp>
 #include <boost/container_hash/hash.hpp>
 #include <caffeine/IR/OperationBase.h>
 #include <fmt/format.h>
@@ -28,297 +26,6 @@ namespace caffeine {
   CAFFEINE_ASSERT((t1) == (t2), fmt::format("{}: {} != {}", (msg), (t1), (t2)))
 #define ASSERT_SAME_TYPES(...)                                                 \
   CAFFEINE_INVOKE_NUMBERED(ASSERT_SAME_TYPES_, __VA_ARGS__)
-
-Operation::Operation() : opcode_(Invalid), type_(Type::void_ty()) {}
-
-Operation::Operation(std::unique_ptr<OperationData>&& data,
-                     std::initializer_list<OpRef> operands)
-    : opcode_(data->opcode()), type_(data->type()), data_(std::move(data)),
-      operands_(operands) {
-  size_t nargs = detail::opcode_nargs(opcode());
-
-  if (nargs != 4) {
-    CAFFEINE_ASSERT(nargs == operands.size(),
-                    fmt::format("invalid number of arguments: {} != {}", nargs,
-                                operands.size()));
-  }
-}
-
-Operation::Operation(Opcode op, Type t, const Inner& inner)
-    : opcode_(static_cast<uint16_t>(op)), type_(t), inner_(inner) {}
-Operation::Operation(Opcode op, Type t, Inner&& inner)
-    : opcode_(static_cast<uint16_t>(op)), type_(t), inner_(std::move(inner)) {}
-
-Operation::Operation(Opcode op, Type t)
-    : opcode_(static_cast<uint16_t>(op)), type_(t), inner_(std::monostate()) {}
-
-Operation::Operation(Opcode op, Type t, const OpRef* operands)
-    : opcode_(static_cast<uint16_t>(op)), type_(t),
-      inner_(OpVec(operands, operands + detail::opcode_nargs(opcode_))) {
-  CAFFEINE_ASSERT(detail::opcode_base(opcode_) != 1,
-                  "Tried to create a constant with operands");
-  // Don't use this constructor to create an invalid opcode.
-  // It'll mess up constructors and destructors.
-  CAFFEINE_ASSERT(op != Invalid);
-  // No opcodes have > 3 operands
-  CAFFEINE_ASSERT(detail::opcode_nargs(opcode_) <= 3, "Invalid opcode");
-}
-
-Operation::Operation(Opcode op, Type t, const OpRef& op0)
-    : opcode_(static_cast<uint16_t>(op)), type_(t), inner_(OpVec{op0}) {
-  CAFFEINE_ASSERT(detail::opcode_base(opcode_) != 1,
-                  "Tried to create a constant with operands");
-  CAFFEINE_ASSERT(detail::opcode_nargs(opcode_) == 1);
-}
-Operation::Operation(Opcode op, Type t, const OpRef& op0, const OpRef& op1)
-    : opcode_(static_cast<uint16_t>(op)), type_(t), inner_(OpVec{op0, op1}) {
-  CAFFEINE_ASSERT(detail::opcode_base(opcode_) != 1,
-                  "Tried to create a constant with operands");
-  CAFFEINE_ASSERT(detail::opcode_nargs(opcode_) == 2);
-}
-Operation::Operation(Opcode op, Type t, const OpRef& op0, const OpRef& op1,
-                     const OpRef& op2)
-    : opcode_(static_cast<uint16_t>(op)), type_(t),
-      inner_(OpVec{op0, op1, op2}) {
-  CAFFEINE_ASSERT(detail::opcode_base(opcode_) != 1,
-                  "Tried to create a constant with operands");
-  CAFFEINE_ASSERT(detail::opcode_nargs(opcode_) == 3);
-}
-
-Operation::Operation(Operation&& op) noexcept
-    : std::enable_shared_from_this<Operation>(), opcode_(op.opcode_),
-      type_(op.type_), inner_(std::move(op.inner_)), data_(std::move(op.data_)),
-      operands_(std::move(op.operands_)) {
-  copy_vtable(op);
-  op.reset();
-}
-
-Operation& Operation::operator=(Operation&& op) noexcept {
-  inner_ = std::move(op.inner_);
-  data_ = std::move(op.data_);
-  operands_ = std::move(op.operands_);
-  type_ = op.type_;
-  opcode_ = op.opcode_;
-
-  copy_vtable(op);
-  op.reset();
-
-  return *this;
-}
-
-void Operation::reset() {
-  opcode_ = Invalid;
-  type_ = Type::void_ty();
-  operands_.clear();
-  data_ = nullptr;
-  inner_ = std::monostate{};
-}
-
-bool Operation::operator==(const Operation& op) const {
-  if (opcode_ != op.opcode_ || type_ != op.type_)
-    return false;
-  if (operands_ != op.operands_)
-    return false;
-
-  if (data_ != op.data_) {
-    if ((bool)data_ != (bool)op.data_)
-      return false;
-
-    if (*data_ != *op.data_)
-      return false;
-  }
-
-  return std::visit(
-      [](const auto& a, const auto& b) {
-        if constexpr (!std::is_same_v<std::decay_t<decltype(a)>,
-                                      std::decay_t<decltype(b)>>) {
-          return false;
-        } else if constexpr (std::is_same_v<std::decay_t<decltype(a)>,
-                                            llvm::APFloat>) {
-          return a.bitwiseIsEqual(b);
-        } else {
-          return a == b;
-        }
-      },
-      inner_, op.inner_);
-}
-bool Operation::operator!=(const Operation& op) const {
-  return !(*this == op);
-}
-
-OpRef Operation::with_new_operands(llvm::ArrayRef<OpRef> operands) const {
-  CAFFEINE_ASSERT(operands.size() == num_operands());
-
-  if (num_operands() == 0)
-    return shared_from_this();
-
-  auto my_operands = std::get<OpVec>(inner_);
-  bool equal = std::equal(std::begin(my_operands), std::end(my_operands),
-                          std::begin(operands), std::end(operands));
-
-  if (equal)
-    return shared_from_this();
-
-  Operation next{(Opcode)opcode(), type(), operands.data()};
-  next.copy_vtable(*this);
-  return constant_fold(std::move(next));
-}
-
-std::string_view Operation::opcode_name() const {
-  return opcode_name(static_cast<Opcode>(opcode()));
-}
-std::string_view Operation::opcode_name(Opcode op) {
-#ifdef MAGIC_ENUM_SUPPORTED
-  std::string_view name = magic_enum::enum_name(op);
-
-  if (name.empty())
-    return "Unknown";
-
-  if (boost::algorithm::starts_with(name, "ICmp"))
-    return "ICmp";
-  if (boost::algorithm::starts_with(name, "FCmp"))
-    return "FCmp";
-
-  return name;
-#else
-  return "<unsupported>";
-#endif
-}
-
-bool Operation::valid() const {
-  return opcode() != 0;
-}
-
-uint16_t Operation::opcode() const {
-  return opcode_;
-}
-
-size_t Operation::num_operands() const {
-  return detail::opcode_nargs(opcode_);
-}
-
-ref<const Operation> Operation::as_ref() const {
-  CAFFEINE_ASSERT(!weak_from_this().expired(),
-                  "Unable to convert non-refcounted Operation "
-                  "instance to a refcounted one");
-  return shared_from_this();
-}
-
-llvm::iterator_range<Operation::const_operand_iterator>
-Operation::operands() const {
-  return llvm::iterator_range<Operation::const_operand_iterator>{
-      const_operand_iterator(this, 0),
-      const_operand_iterator(this, num_operands())};
-}
-
-uint16_t Operation::aux_data() const {
-  return detail::opcode_aux(opcode());
-}
-Type Operation::type() const {
-  return type_;
-}
-
-bool Operation::is_constant() const {
-  return detail::opcode_base(opcode()) == 1;
-}
-
-const Operation& Operation::operator[](size_t idx) const {
-  CAFFEINE_ASSERT(idx < num_operands(),
-                  "Tried to access out-of-bounds operand");
-  return *operand_at(idx);
-}
-
-const OpRef& Operation::operand_at(size_t idx) const {
-  return std::get<OpVec>(inner_)[idx];
-}
-
-template <typename T, typename... Ts>
-static std::ostream& print_spaced(std::ostream& os, const T& first,
-                                  const Ts&... values) {
-  os << first;
-  (..., (os << ' ' << values));
-  return os;
-}
-
-template <typename... Ts>
-static std::ostream& print_cons(std::ostream& os, const Ts&... values) {
-  static_assert(sizeof...(Ts) != 0);
-
-  if constexpr (sizeof...(Ts) == 1) {
-    (os << ... << values);
-    return os;
-  } else {
-    return print_spaced(os << '(', values...) << ')';
-  }
-}
-
-std::ostream& operator<<(std::ostream& os, const Operation& op) {
-  if (const auto* constant = llvm::dyn_cast<Constant>(&op)) {
-    std::string label = fmt::format("const.{}", op.type().bitwidth());
-
-    if (constant->is_named())
-      return print_cons(os, label, constant->name());
-    return print_cons(os, label, constant->number());
-  }
-
-  if (const auto* constant = llvm::dyn_cast<ConstantInt>(&op)) {
-    llvm::SmallString<32> value;
-    constant->value().toString(value, 10, false);
-
-    return print_cons(os, op.type(),
-                      std::string_view(value.data(), value.size()));
-  }
-
-  if (const auto* constant = llvm::dyn_cast<ConstantFloat>(&op)) {
-    return print_cons(os, op.type(), fmt::format("{}", constant->value()));
-  }
-
-  if (const auto* function = llvm::dyn_cast<FunctionObject>(&op)) {
-    return print_cons(os, "function", function->function()->getName().str());
-  }
-
-  std::string name(op.opcode_name());
-  std::transform(name.begin(), name.end(), name.begin(),
-                 [](char c) { return std::tolower(c); });
-
-  if (const auto* icmp = llvm::dyn_cast<ICmpOp>(&op)) {
-    std::string_view cmp = magic_enum::enum_name(icmp->comparison());
-
-    if (cmp.empty())
-      cmp = "<unknown>";
-
-    name.push_back('.');
-    name += cmp;
-  }
-
-  if (const auto* fcmp = llvm::dyn_cast<FCmpOp>(&op)) {
-    std::string_view cmp = magic_enum::enum_name(fcmp->comparison());
-
-    if (cmp.empty())
-      cmp = "<unknown>";
-
-    name.push_back('.');
-    name += cmp;
-  }
-
-  os << '(' << name;
-  size_t nops = op.num_operands();
-  for (size_t i = 0; i < nops; ++i) {
-    os << ' ' << op[i];
-  }
-  return os << ')';
-}
-std::ostream& operator<<(std::ostream& os, Operation::Opcode opcode) {
-  switch (opcode) {
-#define HANDLE_OP(opcode, opname, opclass)                                     \
-  case Operation::opcode:                                                      \
-    return os << #opcode;
-#include "caffeine/IR/Operation.def"
-
-  default:
-    return os << "Unknown(" << (uint16_t)opcode << ")";
-  }
-}
 
 /***************************************************
  * Constant                                        *
